% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{addAnno,annoHub-method}
\alias{addAnno,annoHub-method}
\title{Add an annotation to annoHub}
\usage{
\S4method{addAnno}{annoHub}(annoSource, anAnno, overwrite = TRUE)
}
\arguments{
\item{overwrite}{TRUE to overwrite the existing anno in annoSource, otherwise error is raised when the anno already exists.}
}
\value{
An annoHub object.
}
\description{
addAnno adds a annotation source to the annoHub object.
}
\examples{
library(Mus.musculus)
orgdb=Mus.musculus
ah=new("annoHub")
ah=addAnno(ah, orgdb)
}
