% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{addAnno}
\alias{addAnno}
\title{addAnno generic}
\usage{
addAnno(annoSource, anAnno, ...)
}
\arguments{
\item{annoSource}{An annoHub object.}

\item{anAnno}{An annotation source: txdb=TxDb, ensdb=EnsDb, orgdb=OrganismDb, biomart=Mart, gff=list from useGff(), genes=user provided data.frame.}

\item{...}{dot arguments.}
}
\description{
addAnno is a generic function to add a annotation source to the annoHub object.
}
\seealso{
Other annotation functions: 
\code{\link{annoHub-class}},
\code{\link{getAnnoChrs}()},
\code{\link{getAnnoGenes}()},
\code{\link{getChrs}()},
\code{\link{getGenesRange}()},
\code{\link{isChrConsistent}()},
\code{\link{setDefaultAnno}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
