% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\docType{class}
\name{annoHub-class}
\alias{annoHub-class}
\alias{annoHub}
\title{annoHub object}
\description{
annoHub is an S4 class to store different genome annotation sources.
}
\section{Slots}{

\describe{
\item{\code{annos}}{A list storing different annotation sources: txdb=TxDb, ensdb=EnsDb, orgdb=OrganismDb, biomart=Mart, gff=list from useGff(), genes=data.frame of a user provided gene ranges.}

\item{\code{chrMappings}}{A data frame storing the mappings of chromosome names (seqnames).}

\item{\code{defaultAnno}}{The anno name of the default annotation.}
}}

\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{getAnnoChrs}()},
\code{\link{getAnnoGenes}()},
\code{\link{getChrs}()},
\code{\link{getGenesRange}()},
\code{\link{isChrConsistent}()},
\code{\link{setDefaultAnno}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
