% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{countAPAmarkers}
\alias{countAPAmarkers}
\title{Counts number of APA markers from the result of getAPAmarkers}
\usage{
countAPAmarkers(markerRes, plot = T)
}
\arguments{
\item{markerRes}{A table from getAPAmarkers.}

\item{plot}{TRUE to plot a bar plot, default is TRUE.}
}
\value{
a count table.
}
\description{
countAPAmarkers simply counts and plots number of APA markers from the result of getAPAmarkers.
}
\examples{
data(scPACds)
markerRes=getAPAmarkers(scPACds, group='celltype')
countAPAmarkers(markerRes)
}
\seealso{
Other vizMarkers functions: 
\code{\link{APAindex2PACds}()},
\code{\link{getAPAindexPACds}()},
\code{\link{getAPAmarkers}()},
\code{\link{vizAPAMarkers}()}
}
\concept{vizMarkers functions}
