% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getAnnoChrs,annoHub-method}
\alias{getAnnoChrs,annoHub-method}
\title{Get chrs from annoHub}
\usage{
\S4method{getAnnoChrs}{annoHub}(annoSource, which = NULL)
}
\arguments{
\item{which}{A character of the anno, should be exdb/ensdb/orgdb/gff/biomart/genes.}
}
\value{
A vector of chr names.
}
\description{
getAnnoChrs gets chrs from an anno of annoHub
}
\examples{
library(Mus.musculus)
orgdb=Mus.musculus
annoSource=new("annoHub")
annoSource=addAnno(annoSource, orgdb)
getAnnoChrs(annoSource, 'orgdb')
}
