% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getAnnoChrs}
\alias{getAnnoChrs}
\title{getAnnoChrs generic}
\usage{
getAnnoChrs(annoSource, ...)
}
\arguments{
\item{annoSource}{An annoHub object.}

\item{...}{dot arguments.}
}
\description{
getAnnoChrs is a generic function to get chrs from an anno of annoHub.
}
\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{annoHub-class}},
\code{\link{getAnnoGenes}()},
\code{\link{getChrs}()},
\code{\link{getGenesRange}()},
\code{\link{isChrConsistent}()},
\code{\link{setDefaultAnno}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
