% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getAnnoGenes}
\alias{getAnnoGenes}
\title{Get genomic regions of all genes}
\usage{
getAnnoGenes(annoObj)
}
\arguments{
\item{annoObj}{An annoHub object or any ANNOTYPES.}
}
\value{
A data frame of genomic ranges of all genes, with fixed columns of chr/strand/start/end and other columns starting with 'gene_' like gene_entrez.
}
\description{
getAnnoGenes retrieves genomic ranges of all genes from different genome annotation sources.
}
\examples{
library(Mus.musculus)
orgdb=Mus.musculus
genes=getAnnoGenes(orgdb)
}
\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{annoHub-class}},
\code{\link{getAnnoChrs}()},
\code{\link{getChrs}()},
\code{\link{getGenesRange}()},
\code{\link{isChrConsistent}()},
\code{\link{setDefaultAnno}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
