% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getAPAindexPACds}
\alias{getAPAindexPACds}
\title{Calculate the RUD index for each APA gene in a PACdataset}
\usage{
getAPAindexPACds(PACds, choose2PA = "PD")
}
\arguments{
\item{PACds}{A dataset of the PACdataset class, which should be annotated with genomic regions, e.g., 3UTR.}

\item{choose2PA}{Specify whether and how to choose two PACs. The value can be NULL, PD (choose only proximal and distal sites), farest (choose two PACs that are with the longest distance), most (choose two PACs with the most abundance).}
}
\value{
A new PACdataset with a RUD data frame with rownames denoting genes. Columns are samples, and values are the APA index values.
}
\description{
getAPAindexPACds utilizes movAPA's movAPAindex to calculate the RUD index for each 3'UTR-APA gene in a PACdataset and returns a PACdataset.
This function only implements the RUD index in movAPA, users can use movAPA::movAPAindex for more types of APA index.
}
\examples{
# Calculate the RUD index for each gene.
# Only genes with 3'UTR APA can be used for RUD calculation.
data(scPACds)
iPACds=getAPAindexPACds(scPACds, choose2PA='PD')
movAPA::summary(iPACds)
}
\seealso{
Other vizMarkers functions: 
\code{\link{APAindex2PACds}()},
\code{\link{countAPAmarkers}()},
\code{\link{getAPAmarkers}()},
\code{\link{vizAPAMarkers}()}
}
\concept{vizMarkers functions}
