% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getAPAmarkers}
\alias{getAPAmarkers}
\title{Get APA markers among cell types for single-cell PACdataset}
\usage{
getAPAmarkers(
  PACds,
  group,
  cluster1 = NULL,
  cluster2 = NULL,
  everyPair = TRUE,
  method = c("wilcox"),
  only.pos = FALSE,
  logFC = 0.25,
  min.pct = 0.1
)
}
\arguments{
\item{PACds}{A dataset of the PACdataset class}

\item{group}{The column recording cell categories in PACds@colData.}

\item{cluster1}{A category in the group.}

\item{cluster2}{Another category in the group. If both cluster1/2 are provided, will compare the two clusters.
If only cluster1 or cluster2 is provided, then will compare the given cluster with another group containing all other cells.}

\item{everyPair}{if TRUE and both cluster1/2 are NULL, then will find markers for every pair of clusters in the group.
If cluster1 or cluster2 is provided, then everyPair is automatically set as FALSE.}

\item{method}{Denotes which test to use. See Seurat::FindMarkers(). Available options are:
"wilcox" : Identifies differentially expressed APA genes between two groups of cells using a Wilcoxon Rank Sum test (default)
"t" : Identify differentially expressed APA genes between two groups of cells using the Student's t-test.
"LR" : Uses a logistic regression framework to determine differentially expressed APA genes.
 Constructs a logistic regression model predicting group membership based on each feature individually and compares this to a null model with a likelihood ratio test.
 All the three methods are applicable to the count-type PACdataset, while only wilcox and t are applicable to the ratio-type PACdataset}

\item{only.pos}{Only return positive markers (FALSE by default). If only.pos=FALSE (default), then both positive (avg_log2FC>0) and negative (avg_log2FC<0) markers will be returned. See Seurat::FindMarkers().}

\item{logFC}{log2FC threshold to filter significant markers, default is 0.25.}

\item{min.pct}{Only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations.
Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.1. See Seurat::FindMarkers().}
}
\value{
A data frame of APA marker list with columns: p_val,avg_log2FC,pct.1,pct.2,p_val_adj,cluster1,cluster2,rowid,direction.
\itemize{
\item{p_val}: p value.
\item{avg_log2FC}: log fold-chage of the average expression or APA ratio (depends on the PACds) between the two groups. Positive values indicate that the gene or pA is more highly expressed in the first group.
\item{pct.1}: The percentage of cells where the gene or pA is detected in the first group.
\item{pct.2}: The percentage of cells where the gene or pA is detected in the second group.
\item{p_val_adj}: adjusted p value.
\item{rowid}: the row name of the marker in PACds@counts, normally is a gene name or id.
\item{direction.}: positive if avg_log2FC>0, otherwise negative.
}
}
\description{
getAPAmarkers finds APA markers among different cell types, leveraging Seurat's FindMarkers and FindAllMarkers functions on a PACdataset.
}
\examples{
data(scPACds)
getAPAmarkers(scPACds, group='celltype')
# compare ES to other groups, and only return positive markers,
# those with higher score in ES (avg_log2FC>0)
getAPAmarkers(scPACds, group='celltype', cluster1='ES',
             only.pos=TRUE, logFC=0.15, min.pct=0.1)
# using default parameters, only.pos=F, logFC=0.25, min.pct=0.1
getAPAmarkers(scPACds, group='celltype', cluster1='ES')
# compare two cell types
getAPAmarkers(scPACds, group='celltype', cluster1='ES', cluster2='RS')
# statistical test used for marker detection
getAPAmarkers(scPACds,  group='celltype',
              cluster1='ES', cluster2='RS', method='t')
getAPAmarkers(scPACds,  group='celltype',
             cluster1='ES', cluster2='RS', method='LR')
}
\seealso{
Other vizMarkers functions: 
\code{\link{APAindex2PACds}()},
\code{\link{countAPAmarkers}()},
\code{\link{getAPAindexPACds}()},
\code{\link{vizAPAMarkers}()}
}
\concept{vizMarkers functions}
