% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getChrs}
\alias{getChrs}
\title{Get chromosome names}
\usage{
getChrs(obj, which = NULL)
}
\arguments{
\item{obj}{An obj of any class (PACdataset, bam file list, annoSource list, txdb/gff/ensdb/biomart/orgdb, dataframe/matrix, bsgenome, fafile).}

\item{which}{Specify the element in a list for retrieving chr names or a column in a dataframe or matrix denoting the chr column. For a list (such as annoSource), the obj in the list to get chr is specified by which. For a dataframe/matrix, which specifies the chr column name, when 'seqnames' or 'chr' does not exist in colnames.
By default, if which=NULL, then the first item of a list (if obj is a list) is considered, or 'seqnames' or 'chr' column is searched first (when obj is matrix/data.frame).}
}
\value{
A character vector with chromosome names.
}
\description{
getChrs gets the chromosome names of any given object
}
\examples{
getChrs(1:3)
# get chrs from a dataframe or matrix
data(scPACds)
getChrs(scPACds@anno)
getChrs(scPACds@anno, which='seqnames') ##error: which not in df
getChrs(scPACds@anno, which='chr')
# get chrs from a PACdataset
getChrs(scPACds)
\dontrun{
# get chr names from a parse GFF object
gff <- useGff(gff ="data/Arabidopsis_thaliana.TAIR10.49.gtf")
getChrs(gff)
}
# get chrs from a TxDb object
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
getChrs(TxDb.Hsapiens.UCSC.hg19.knownGene)
# get chrs from an EnsDb object
library(EnsDb.Hsapiens.v75)
getChrs(EnsDb.Hsapiens.v75)
# get chrs from an OrgDb object
library(Homo.sapiens)
getChrs(Homo.sapiens)
# get chrs from biomart
bm <- useMart(biomart="plants_mart",
             host="https://plants.ensembl.org", dataset="athaliana_eg_gene")
getChrs(bm)
# different annotation sources can be put in a list,
# which=NULL will get chrs from the first element in the list
getChrs(obj=list(biomart=bm), which=NULL)
# specify which
getChrs(obj=list(biomart=bm,
       txdb=TxDb.Hsapiens.UCSC.hg19.knownGene),
       which='txdb')
\dontrun{
# get chrs from a BAM file list
bams=readBAMFileNames(...)
getChrs(bams)
}
}
\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{annoHub-class}},
\code{\link{getAnnoChrs}()},
\code{\link{getAnnoGenes}()},
\code{\link{getGenesRange}()},
\code{\link{isChrConsistent}()},
\code{\link{setDefaultAnno}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
