% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getGenesRange}
\alias{getGenesRange}
\title{Get genomic regions of given genes}
\usage{
getGenesRange(genes, annoObj, rt = c("gr", "str", "list"))
}
\arguments{
\item{genes}{A vector of gene names/symbols/ids, which could be mixed with names/symbols/ids for some annotation sources.
\itemize{
\item{TxDb}: 'genes' could only be entrezid (e.g., 9991).
\item{EnsDb}: 'genes' could be geneID (e.g., ENSG00000121410), gene_name (e.g., NGA3), symbol (e.g., NGA3), entrezid (e.g., 9991), but mixed is not allowed and the provided list of genes would be searched in above order.
\item{OrgDb}: 'genes' could be GENENAME (e.g., NGA3),SYMBOL (e.g., NGA3), ENSEMBL (e.g., ENSG00000121410), GENEID (e.g., 9991), or mixed.
\item{BioMart}: 'genes' could be ensembl_gene_id (e.g., ENSG00000121410), external_gene_name (e.g., NGA3), entrezgene_id (e.g., 9991), but mixed is not allowed and the provided list of genes would be searched in above order.
\item{genes}:  `genes` could be mixed, and will consider all columns with names containing 'gene' as gene columns.
\item{gff}: 'genes' could be gene_id (e.g., AT1G03987), gene_name (e.g., NGA3), or mixed.
}}

\item{annoObj}{An annoHub object or any ANNOTYPES.}

\item{rt}{Specify the output format, GR to output GRanges or other to output a string like 'chr:strand:start:end'. When output a string, only the first row of GRanges will be returned.}
}
\value{
A GRanges object or a string like 'chr:strand:start:end'.
}
\description{
getGenesRange retrieves genomic ranges of given genes from different genome annotation sources.
}
\examples{
\dontrun{
# get gene ranges from a GFF object
gff <- useGff(gff ="data/Arabidopsis_thaliana.TAIR10.49.gtf")
getGenesRange(genes=c('AT1G03987', 'NGA3'),
             annoSource=new('annoHub', annos=list(gff=gff)))
# from a TxDb object
getGenesRange(genes=c('ENSG00000121410', 'cadherin 2', 'A1BG', '9991'),
       new('annoHub', annos=list(txdb=TxDb.Hsapiens.UCSC.hg19.knownGene)))
# from an EnsDb
getGenesRange(genes=c('ENSG00000121410', 'cadherin 2', 'A1BG', '9991'),
                        new('annoHub', annos=(ensdb=EnsDb.Hsapiens.v75)))
# from a OrgDb
getGenesRange(genes=c('ENSG00000121410', 'cadherin 2', 'A1BG', '9991'),
              new('annoHub', annos=(orgdb=Homo.sapiens)))
# from a Biomart object
bm <- useMart(biomart="plants_mart", host="https://plants.ensembl.org",
              dataset="athaliana_eg_gene")
getGenesRange(genes=c('AT1G03987', 'cadherin 2', 'A1BG', '9991'),
              new('annoHub', annos=(biomart=bm)))
# return a string
genesGR=getGenesRange(gene='ENSG00000121410',
                 new('annoHub', annos=list(orgdb=Homo.sapiens), rt='str'))
}
}
\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{annoHub-class}},
\code{\link{getAnnoChrs}()},
\code{\link{getAnnoGenes}()},
\code{\link{getChrs}()},
\code{\link{isChrConsistent}()},
\code{\link{setDefaultAnno}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
