% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getTrackBams}
\alias{getTrackBams}
\title{Get the BAM track}
\usage{
getTrackBams(bams, genomicRegion, vizTheme = NULL)
}
\arguments{
\item{bams}{A dataframe describing the BAM file list. See the output of readBAMFileNames().}

\item{genomicRegion}{A string with four fields describing a genomic region, like 1:-:1000:100000. The BAM coverages on the same strand in the BAM file will be extracted.}

\item{vizTheme}{a vizTHEME-like list to specify the theme of the track.}
}
\value{
Tracks of BAM coverages. This function first subsets BAM coverages in the given genomic region and returns the track in different styles based on vizTheme.
        The number of tracks depends on vizTheme. The tracks can be individual conditions (cell types), merged data (AVG or SUM), or collapsed coverages. The colors of bams set in vizTheme$bams.col are as the order in the bams list.
}
\description{
getTrackBams gets the plot track(s) for a BAM file list.
}
\examples{
\dontrun{
tk=getTrackBams(bams, genomicRegion='1:+:1000:10000', vizTheme=vizTheme)
}
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackCells}()},
\code{\link{getTrackGeneModel}()},
\code{\link{getTrackPACds}()},
\code{\link{readBAMFileNames}()},
\code{\link{setVizTheme}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracksOfRegion}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
