% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getTrackCells}
\alias{getTrackCells}
\title{Get the track of pA in individual cells}
\usage{
getTrackCells(
  PACds,
  genomicRegion = NULL,
  gene = NULL,
  PAs = NULL,
  group = NULL,
  sortMethod = c("sum", "diffusion"),
  sortCells = c("none", "all", "group"),
  log = FALSE,
  PA.columns = "coord",
  PA.width = 100,
  annoCoord = NULL,
  title = "cells",
  vizTheme = NULL
)
}
\arguments{
\item{PACds}{A PACdataset object.}

\item{genomicRegion}{A string with four fields describing a genomic region, like 1:-:1000:100000.}

\item{gene}{A gene name corresponding to the gene column in PACds@anno.}

\item{PAs}{A vector of pA ids (rownames of PACds@counts).}

\item{group}{The column name recording cell categories in PACds@colData.}

\item{sortMethod}{To set the method for ordering cells, sum (using the sum of all pAs) (default), or diffusion (using diffusion map to order cells -- see Millefy Package).}

\item{sortCells}{To sort cells within group (group) or without group (all), or not sort (none, default).}

\item{log}{TRUE to log2(1+PACds@counts) for count-type PACds.}

\item{PA.columns}{A string denoting the pA coordinates (e.g., coord) or ranges (e.g., UPA_start:UPA_end).}

\item{PA.width}{If PA.columns is a position (e.g., coord), then the pA range can be expanded by PA.width to be visualized more clearly in the track plot.
For example, if PA.width=10, then the pA position would be converted to a range of 2*PA.width+1 with the pA in the center.}

\item{annoCoord}{The X-position to put the annotation bar of cell groups, only used when group is not NULL.}

\item{title}{Title of the track.}

\item{vizTheme}{a vizTHEME-like list to specify the theme of the track.}
}
\value{
One track to show pA counts or ratios in individual cells. The bar width of each pA depends on PA.columns and/or PA.width. If PA.columns is like 'coord' then the bar width is 2*PA.width+1; if PA.columns is like 'start:end', then the bar width of a pA is end-start+1 of the respective pA.
}
\description{
getTrackCells gets the track of pA/gene ratios or counts in individual cells
}
\examples{
\dontrun{
sc.tk1=getTrackCells(PACds,gene=gene,
group='celltype', sortMethod='sum', sortCells='group',log=TRUE,
PA.columns="coord", PA.width=100,title='diffs')
}
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackGeneModel}()},
\code{\link{getTrackPACds}()},
\code{\link{readBAMFileNames}()},
\code{\link{setVizTheme}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracksOfRegion}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
