% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getTrackGeneModel}
\alias{getTrackGeneModel}
\title{Get gene model track}
\usage{
getTrackGeneModel(
  genomicRegion = NULL,
  gene = NULL,
  annoSource = new("annoHub"),
  title = "genemodel",
  vizTheme = NULL
)
}
\arguments{
\item{genomicRegion}{A string with four fields describing a genomic region, like 1:-:1000:100000'.}

\item{gene}{A gene name, gene symbol, or gene id'.}

\item{annoSource}{An annoHub object, the defaultAnno would be used to get gene models. If is.empty, then use the genomicRegion as axis.}

\item{title}{Title of the track, default is genemodel.}

\item{vizTheme}{A vizTHEME-like list to specify the theme of the track.}
}
\value{
A track of the gene model.
}
\description{
getTrackGeneModel retrieves a gene model track from a given gene or a genomic range, based on different annotation sources.
}
\examples{
\dontrun{
# get track from a genomic region
gff <- useGff(gff ="data/Arabidopsis_thaliana.TAIR10.49.gtf")
getTrackGeneModel(genomicRegion='1:-:1000:100000',
               annoSource=new("annoHub", annos=list(gff=gff)))
getTrackGeneModel(genomicRegion='chr1:-:1000:100000',
                  annoSource=new("annoHub",
                      annos=list(txdb=TxDb.Hsapiens.UCSC.hg19.knownGene)))
getTrackGeneModel(genomicRegion='1:-:1000:100000',
           annoSource=new("annoHub", annos=list(ensdb=EnsDb.Hsapiens.v75)))
getTrackGeneModel(genomicRegion='chr1:-:1000:100000',
         annoSource=new("annoHub", annos=list(orgdb=Homo.sapiens)))
bm <- useMart(biomart="plants_mart",
          host="https://plants.ensembl.org", dataset="athaliana_eg_gene")
getTrackGeneModel(genomicRegion='1:+:500000:520000',
                  annoSource=new("annoHub", annos=list(biomart=bm)))
# get track from a gene
getTrackGeneModel(gene='AT1G01200',
                     annoSource=new("annoHub", annos=list(biomart=bm)))
# we can also get the genomic range of a gene first
gr=getGenesRange(genes=c('AT1G01200'),
                annoSource=new("annoHub", annos=list(biomart=bm), rt='str'))
# and then get the gene model track
getTrackGeneModel(genomicRegion=gr,
                 annoSource=new("annoHub", annos=list(biomart=bm)))
}
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackCells}()},
\code{\link{getTrackPACds}()},
\code{\link{readBAMFileNames}()},
\code{\link{setVizTheme}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracksOfRegion}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
