% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{getTrackPACds}
\alias{getTrackPACds}
\title{Get the pA track}
\usage{
getTrackPACds(
  PACds,
  genomicRegion = NULL,
  gene = NULL,
  PAs = NULL,
  PA.show = c("pos"),
  log = FALSE,
  PA.columns = "coord",
  PA.width = NULL,
  title = NULL,
  vizTheme = NULL
)
}
\arguments{
\item{PACds}{A PACdataset object.}

\item{genomicRegion}{A string with four fields describing a genomic region, like 1:-:1000:100000.}

\item{gene}{A gene name corresponding to the gene column in PACds@anno.}

\item{PAs}{A vector of pA ids (rownames of PACds@counts).}

\item{PA.show}{A vector describing the information of pA to extract, e.g., PA.show=c('pos','tag').
PA.show=pos means extracting the coordinates of pAs; other values (e.g., tag, tot_tagnum) should be the column in PACds@anno or PACds@counts denoting the expression level of pAs.}

\item{log}{TRUE to log2(PACds@counts+1) applicable when PA.show to show tag not pos.}

\item{PA.columns}{A string denoting the pA coordinates (e.g., coord) or ranges (e.g., UPA_start:UPA_end).}

\item{PA.width}{If PA.columns is a position (e.g., coord), then the pA range can be expanded by PA.width to be visualized more clearly in the track plot.
For example, if PA.width=10, then the pA position would be converted to a range of 2*PA.width+1 with the pA in the center.}

\item{title}{Title for the track, should be the same length as PA.show.}

\item{vizTheme}{a vizTHEME-like list to specify the theme of the track.}
}
\value{
One or more tracks of pAs This function first subsets PACds by genomie region or gene name or pA ids, and retrieves the pA information (coordinates and/or expression values) according to PA.show. The number of tracks is equal to the length of PA.show.
        The bar width of each pA depends on PA.columns and/or PA.width. If PA.columns is like 'coord' then the bar width is 2*PA.width+1; if PA.columns is like 'start:end', then the bar width of a pA is end-start+1 of the respective pA.
}
\description{
getTrackPACds gets the plot track(s) for a PACdataset.
}
\examples{
data(scPACds)
tk=getTrackPACds(scPACds, genomicRegion='1:+:1000:10000', PA.show=c("pos","tag"),
             PA.columns="coord", PA.width=10)
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackCells}()},
\code{\link{getTrackGeneModel}()},
\code{\link{readBAMFileNames}()},
\code{\link{setVizTheme}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracksOfRegion}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
