% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{isChrConsistent}
\alias{isChrConsistent}
\title{Check consistency of chromosome names among different objects}
\usage{
isChrConsistent(..., which = NULL, exact = FALSE, verbose = TRUE)
}
\arguments{
\item{...}{Any number of objects, e.g, PACdataset, bam file names, an annoHub or a parsed gff, txdb/gff/ensdb/biomart/orgdb, dataframe/matrix, bsgenome, fafile. See getChrs() for supported classes.}

\item{which}{Specify the element in a list or a column in a dataframe or matrix for retrieving chr names.
For an annoHub, the object is specified by which. For a dataframe/matrix, 'which' specifies the chr column name (if seqnames or chr not exists).}

\item{exact}{Define the consistency of chr names among the given objects, with FALSE being partially overlapped of chrs (>=1 intersects)  or TRUE completely the same.}

\item{verbose}{TRUE to print the message.}
}
\value{
TRUE if the chr names are (partially or exactly) consistent among objects, or FALSE if not.
}
\description{
isChrConsistent checks the consistency of chromosome names among any number of given objects, e.g., PACdataset, dataframe/matrix, genome annotation sources, BSgenome.
}
\examples{
\dontrun{
# Load different genome annotation sources
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(EnsDb.Hsapiens.v75)
txdb=TxDb.Hsapiens.UCSC.hg19.knownGene
ensdb=EnsDb.Hsapiens.v75
orgdb=Homo.sapiens
bm <- useMart(biomart="plants_mart", host="https://plants.ensembl.org", dataset="athaliana_eg_gene")
gff <- useGff(gff ="data/Arabidopsis_thaliana.TAIR10.49.gtf")
# check the consistency of chr names
isChrConsistent(txdb, ensdb, orgdb) #FALSE
isChrConsistent(orgdb, txdb) #TRUE
isChrConsistent(orgdb, txdb, exact=TRUE) #TRUE
isChrConsistent(gff, bm, exact=TRUE) # #TRUE
isChrConsistent(c(1:5,'Mt','Pt'), bm, exact=TRUE) #TRUE
isChrConsistent(c(1:5,'chr1','chr2'), bm, exact=TRUE) ##FALSE
isChrConsistent(c(1:5,'chr1','chr2'), bm, exact=FALSE) ##TRUE
ah=new("annoHub", annos=list(biomart=bm, gff=gff))
isChrConsistent(c(1:5,'chr1','chr2'), ah) ##TRUE
isChrConsistent(c(1:5,'chr1','chr2'), ah, which='gff') ##TRUE
isChrConsistent(c(1:5,'chr1','chr2'), ah, which='XXX') ##Error!
isChrConsistent(c(1:5,'chr1','chr2'), ah, bams, which='gff') ##TRUE
isChrConsistent(c(1:5,'chr1','chr2'), ah, bams, which='gff', exact=TRUE) ##FALSE
}
}
\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{annoHub-class}},
\code{\link{getAnnoChrs}()},
\code{\link{getAnnoGenes}()},
\code{\link{getChrs}()},
\code{\link{getGenesRange}()},
\code{\link{setDefaultAnno}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
