% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{logPACds}
\alias{logPACds}
\title{Log transformation of a PACdataset
logPACds does log2/10 transformation of a PACdataset.}
\usage{
logPACds(PACds, log = 2, psu = 1, logWhenCounts = TRUE)
}
\arguments{
\item{PACds}{A PACdataset.}

\item{log}{Can be 2 or 10 means log2 or log10.}

\item{psu}{Pseudo count to avoid log(0).}

\item{logWhenCounts}{If TRUE then only do log when PACds's dataType is count.}
}
\value{
A PACdataset.
}
\description{
Log transformation of a PACdataset
logPACds does log2/10 transformation of a PACdataset.
}
\examples{
logPACds(PACds)
}
\concept{PACdataset functions}
