% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{readBAMFileNames}
\alias{readBAMFileNames}
\title{Unify the input BAM file names}
\usage{
readBAMFileNames(
  bam.files,
  bam.path = NULL,
  bam.labels = NULL,
  bam.groups = NULL
)
}
\arguments{
\item{bam.files}{A >=2 columns data frame with file name information (the first column is file name, the second is strand, third is label -- optional, the fourth is group -- optional),
or a list of file name(s) with or without file path.}

\item{bam.path}{File path of the bam.files.}

\item{bam.labels}{Short labels for each file of bam.files, default is BAM1...BAMn.}

\item{bam.groups}{Groups for bam.files, default is 'g1' for all bam.files.}
}
\value{
A data frame storing BAM file information with fileName, group, label. This function will check the existence of each file (and the corresponding .bai file)!
}
\description{
readBAMFileNames loads BAM file names and checks the validity.
}
\examples{
## way1: load bam file names from a dataframe
## suppose there is a dataframe with three columns (filename, label, group)
# create three temporary bam files
file.create(c('f1.bam','f2.bam','c3.bam','f1.bam.bai',
              'f2.bam.bai','c3.bam.bai'))
bam.files=cbind(c('f1.bam','f2.bam','c3.bam'),
               c('ctrl1','ctrl2','treat1'),
               g=c('ctrl','ctrl','treat'))
readBAMFileNames(bam.files)
## only provide filename, strand and label,
## then the group is "g1" for all bam files
readBAMFileNames(bam.files[, 1:3])
## way2: load bam file names from a file name list,
## with all files in the current path
readBAMFileNames(bam.files=c('f1.bam','f2.bam','c3.bam'),
                 bam.labels=c('ctrl1','ctrl2','treat1'),
                 bam.groups=c('ctrl','ctrl','treat'))
# delete the created temporary bam files
file.remove(c('f1.bam','f2.bam','c3.bam','f1.bam.bai',
             'f2.bam.bai','c3.bam.bai'))
## way3: load bam file names in a given Path
## create 2 temporary bam files in one path
dir.create('_dir1')
file.create('_dir1/bam1.bam','_dir1/bam2.bam',
           '_dir1/bam1.bam.bai','_dir1/bam2.bam.bai')
readBAMFileNames(bam.files=c('bam1.bam', 'bam2.bam'),
                bam.path='_dir1')
file.remove(c('_dir1/bam1.bam','_dir1/bam2.bam',
            '_dir1/bam1.bam.bai','_dir1/bam2.bam.bai', '_dir1'))
## way4: load bam file names in different paths
## create two temporary bam files in different paths
dir.create('_dir1'); dir.create('_dir2')
file.create('_dir1/bam1.bam','_dir1/bam1.bam.bai',
            '_dir2/bam2.bam' , '_dir2/bam2.bam.bai')
readBAMFileNames(bam.files=c('_dir1/bam1.bam', '_dir2/bam2.bam'),
                bam.labels=c('ctrl1','ctrl2'), bam.groups=c('ctrl','ctrl'))
file.remove('_dir1/bam1.bam','_dir1/bam1.bam.bai', '_dir2/bam2.bam' ,
            '_dir2/bam2.bam.bai', '_dir1','_dir2')
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackCells}()},
\code{\link{getTrackGeneModel}()},
\code{\link{getTrackPACds}()},
\code{\link{setVizTheme}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracksOfRegion}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
