% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{reduceDim}
\alias{reduceDim}
\title{Reduce dimensions by running the Seurat's pipeline on a PACdataset}
\usage{
reduceDim(PACds, nfeature = 2000, dims = 1:10, dimLabel = "umap", norm = FALSE)
}
\arguments{
\item{PACds}{A dataset of the PACdataset class.}

\item{nfeature}{Number of features to select as top variable features, used in the 'vst' method.}

\item{dims}{Which dimensions from PCA to use as input features .}

\item{dimLabel}{The label added or overwritten in the colData of PACds.}

\item{norm}{If norm=TRUE and the datatype of PACds is count, then LogNormalize will be applied.}
}
\value{
A PACdataset with extra columns denoting the 2D-embeddings of the UMAP plot.
}
\description{
reduceDim reduces dimensions by running the Seurat's pipeline on a PACdataset.
}
\examples{
data(scPACds)
## get embeddings using the pA count matrix with normalization
PACds=reduceDim(scPACds, nfeature=2000, dims=1:10, dimLabel='umap_norm', norm=TRUE)
## without normalization
PACds=reduceDim(scPACds, nfeature=2000, dims=1:10, dimLabel='umap_raw', norm=FALSE)
}
\seealso{
Other vizUMAP functions: 
\code{\link{vizUMAP}()}
}
\concept{vizUMAP functions}
