% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\docType{data}
\name{scPACds}
\alias{scPACds}
\alias{.asDf}
\title{A demo PACdataset of mouse sperm cells}
\format{
A PACdataset with 974 PACs from mouse sperm cells.
}
\source{
Generated by scAPAtrap and annotated by movAPA.
}
\usage{
.asDf(dat)
}
\description{
A dataset of the PACdataset class from mouse sperm cells, containing 974 PACs from 413 genes.
There are total 955 cells from three cell types (SC, Spermatocytes; RS, Round spermatids; ES, Elongating spermatids).
This dataset contains the gene Odf4 (252868, ENSMUSG00000032921) and Arl2bp (107566, ENSMUSG00000031776) presented in Wu et al, Briefings Bioinf, 2021.
The raw data was obtained from NCBI (accession number: GSM280334), and pAs were extracted by scAPAtrap.
Only genes containing at least one top 500 pAs ranked by the total number of reads were retained in this dataset for demonstration.
}
\examples{
data(scPACds)
}
