% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{setDefaultAnno,annoHub-method}
\alias{setDefaultAnno,annoHub-method}
\title{Set a default anno for annoHub}
\usage{
\S4method{setDefaultAnno}{annoHub}(annoSource, annoName)
}
\value{
An annoHub object.
}
\description{
setDefaultAnno sets the default annotation source for an annoHub.
}
\examples{
library(Mus.musculus)
orgdb=Mus.musculus
annoSource=new("annoHub")
annoSource=addAnno(annoSource, orgdb)
annoSource=setDefaultAnno(annoSource, 'orgdb')
annoSource=setDefaultAnno(annoSource, 1)
}
