% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{setDefaultAnno}
\alias{setDefaultAnno}
\title{setDefaultAnno generic}
\usage{
setDefaultAnno(annoSource, annoName)
}
\arguments{
\item{annoSource}{An annoHub object.}

\item{annoName}{A character or a numeric index of the anno.
If it is a character, should be exdb/ensdb/orgdb/gff/biomart/genes.}
}
\description{
setDefaultAnno is a generic function to set the default annotation source for an annoHub.
}
\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{annoHub-class}},
\code{\link{getAnnoChrs}()},
\code{\link{getAnnoGenes}()},
\code{\link{getChrs}()},
\code{\link{getGenesRange}()},
\code{\link{isChrConsistent}()},
\code{\link{useGff}()}
}
\concept{annotation functions}
