% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{setStatTheme}
\alias{setStatTheme}
\title{Set the plot theme for vizStats}
\usage{
setStatTheme(statTheme, check = TRUE)
}
\arguments{
\item{statTheme}{A list similar to statTHEME, containing the following parameters.
\itemize{
\item{group.cols}: parameter for vizStats(box/violin/point), colors for cell groups, default is `c(brewer.pal(8, "Set2"), brewer.pal(8, "Set1"))`.
\item{outlier.colour}: parameter for vizStats(box/violin/point), color for outliers, defaut is "grey".
\item{outlier.shape}: parameter for vizStats(box/violin/point), shape for outliers, defaut is 8.
\item{outlier.size}: parameter for vizStats(box/violin/point), size for outliers, default is 1.
\item{alpha}: parameter for vizStats(box/violin/point), the transparent alpha, default is 0.8.

\item{scale.low.col}: parameter for bubble/markers-dot/heatmap, low point of the gradients, default is '#8d9fbb'.
\item{scale.high.col}: parameter for bubble/markers-dot/heatmap, high point of the gradients, default is '#aa3e53',


\item{umap.point.size}: parameter for the UMAP plot, the point size, default is 1.
\item{umap.legend.size}: parameter for the UMAP plot, the legend text size, default is 6.
\item{umap.legend.key.size}: parameter for the UMAP plot, the legend key size, default is 0.4.
\item{umap.legend.point.size}: parameter for the UMAP plot, the legend point size, default is 1.5.

\item{xgroup}: parameter for the marker plots (iolin/dot plots), whether show cell groups on X-axis, default is TRUE.

\item{hm.disp.min}: parameter for the heatmap, default is -2.5, details see Seurat::DoHeatmap.
\item{hm.disp.max}: Default is NULL, details see Seurat::DoHeatmap.
\item{hm.label }:  Default is TRUE, details see Seurat::DoHeatmap.
\item{hm.size }:  Default is 5.5, details see Seurat::DoHeatmap.
\item{hm.hjust }:  Default is 0, details see Seurat::DoHeatmap.
\item{hm.angle }:  Default is 45, details see Seurat::DoHeatmap.
\item{hm.raster }:  Default is TRUE, details see Seurat::DoHeatmap.
\item{hm.draw.lines }: Default is TRUE, details see Seurat::DoHeatmap.
\item{hm.lines.width }: Default is NULL, details see Seurat::DoHeatmap.
\item{hm.group.bar.height }: Default is 0.02, details see Seurat::DoHeatmap.

\item{dot.scale}: parameter for the dot plot, default is FALSE.

\item{umaps.ncol}: parameter for the markers' multiple UMAP plots, which set the number of columns in the plot, default is 2.
}}

\item{check}{Whether to check the validity of values (default is TRUE).}
}
\value{
A list.
}
\description{
setStatTheme sets the drawing theme for vizStats, by setting default values using statTHEME and checking the validity.
}
\examples{
setStatTheme(NULL)
setStatTheme(list(group.cols=brewer.pal(8, "Set1")), check=TRUE)
}
\seealso{
Other vizStats functions: 
\code{\link{vizStats}()}
}
\concept{vizStats functions}
