% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{setVizTheme}
\alias{setVizTheme}
\title{Set the plot theme for vizTracks}
\usage{
setVizTheme(vizTheme, check = TRUE)
}
\arguments{
\item{vizTheme}{A list similar to vizTHEME, containing the following options for plot.
\itemize{
\item{pals}: colors for stat plots, default is `brewer.pal(8, "Set2")`. Normally denotes the colors for each cell types/groups.

\item{bam.covMerge}: for bam coverage plot, could be no/avg/sum, default is `no`. Define how to merge coverages from different cell groups.
\item{bam.showMerge}: for bam coverage plot, default is FALSE. Determine whether to show the bam coverage.
\item{bam.showSingle}: for bam coverage plot, default is TRUE. Determine whether to show each cell group in an individual track.
\item{bam.covStyle}: for bam coverage plot, could be cov/curve, default is `cov`. `cov` means plot coverages like area under a line, `curve` means line.
\item{bam.collapse}: for bam coverage plot, default is FALSE. Determine whether to collapse areas or lines of individual cell groups into one track.

\item{bams.log2}: for bam coverage data presentation, default is FALSE, which means not to log2 the coverage value.
\item{bams.ysame}: for bam coverage data presentation, default is TRUE, which means set all Y-axis limit of different bam coverage tracks the same.
\item{bams.ylims}: for bam coverage data presentation, default is NULL. It can be set as `c(10, 100)` for all bam coverage tracks' Y-axis.

\item{bams.col}: line colors for individual bams, default is `brewer.pal(8, "Set2")`.
\item{bams.fill}: filled colors for individual bams, default is, default is `brewer.pal(8, "Set2")`.


\item{bam.covMerge.col}: line color for merged bam, default is "brown".
\item{bam.covMerge.fill}: filled color for merged bam, default is "brown".

\item{PA.col}: line color for a pA site, default is "#cf9198".
\item{PA.fill}: filled color for a pA site, default is "#cf9198".
\item{PA.shape}: shape for a pA site, default is `-1` (line), set 0-25 for other shapes used in geom_point.
\item{PA.shape.size}: size for the pA site when PA.shape is not -1, default is 3.

\item{gm.utr5.fill}: colors for the gene model's 5'UTR, default is "black".
\item{gm.utr3.fill}: colors for the gene model's 3'UTR, default is"grey".
\item{gm.exon.fill}: colors for the exon regions, default is "brown".
\item{gm.line}: colors for the lines (intron), default is "black"/

\item{cells.group.cols}: colors for the single-cell plot by vizCells, default is `c(brewer.pal(8, "Set2"), brewer.pal(8, "Set1"))`.
\item{cells.scale.low}: colors for low ends of the gradient, default is 'white',
\item{cells.scale.high}: colors for high ends of the gradient, default is 'black'.
\item{cells.scale.mid }: colors for mid point of the gradient, default is "darkorange".
\item{cells.annoBar.pos}: position for the annotation bar of cell groups, could be left (default)/right/mid.
}}

\item{check}{Whether to check the validity of values (default is TRUE).}
}
\value{
A list.
}
\description{
setVizTheme sets the drawing theme for vizTracks, by setting default values in vizTHEME and checking the validity.
}
\examples{
## print the default vizTHEME
setVizTheme(NULL)
## modify a parameter, other parameters would be the default
setVizTheme(list(pals=brewer.pal(8, "Set1")), check=TRUE)
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackCells}()},
\code{\link{getTrackGeneModel}()},
\code{\link{getTrackPACds}()},
\code{\link{readBAMFileNames}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracksOfRegion}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
