% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{subsetPACds2GR}
\alias{subsetPACds2GR}
\title{Subset a PACdataset and convert to GRanges}
\usage{
subsetPACds2GR(
  PACds,
  genomicRegion = NULL,
  gene = NULL,
  PAs = NULL,
  PA.columns = "coord",
  PA.width = 10,
  mcols = NULL
)
}
\arguments{
\item{PACds}{A PACdataset object.}

\item{genomicRegion}{A string with four fields describing a genomic region, like 1:-:1000:100000.}

\item{gene}{A gene name corresponding to the gene column in PACds@anno.}

\item{PAs}{A vector of pA ids corresponding to the rownames of PACds@counts.}

\item{PA.columns}{A string denoting the column name of the pA coordinates (e.g., coord) or the pA ranges (e.g., UPA_start:UPA_end).}

\item{PA.width}{If PA.columns is a position (e.g., coord), then the pA range can be expanded by PA.width to be visualized more clearly in the track plot.
For example, if PA.width=10, then the pA position would be converted to a range of 2*PA.width+1 with the pA in the center.}

\item{mcols}{Extra columns from PACds@anno or PACds@counts to keep as mcols in the final GRange object.
This could be useful when extracting expression values of the PACds for the track plot.}
}
\value{
A GRanges object contains extra columns of coord (the end of a pA range if PA.columns like start:end) and mcols.
}
\description{
subsetPACds2GR subsets a PACdataset by a genomic region, gene id/name/symbol, or pA ids, and converts the result into GRanges.
}
\examples{
# subset PACds in the given region and
# add the "tot_tagnum" column (should be in PACds@anno or PACds@counts) to the returned GR object.
data(scPACds)
subsetPACds2GR(scPACds, genomicRegion='chr1:+:1000:10000', mcols=c('tot_tagnum'))
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackCells}()},
\code{\link{getTrackGeneModel}()},
\code{\link{getTrackPACds}()},
\code{\link{readBAMFileNames}()},
\code{\link{setVizTheme}()},
\code{\link{vizTracksOfRegion}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
