% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{useGff}
\alias{useGff}
\title{Parse a genome annotation file or a TxDb object}
\usage{
useGff(gff = NULL, txdb = NULL)
}
\arguments{
\item{gff}{An annotation file in gtf/gff3 format.}

\item{txdb}{A TxDb object, e.g., TxDb.Athaliana.BioMart.plantsmart28.
Either gff or txdb should be provided.}
}
\value{
A parsed genome annotation stored in a list with three dataframes.
}
\description{
useGff parse genome annotation file of gtf/gff3 format or genome annotation object of TxDb, which could be further used for annotating a PACdataset.
}
\examples{
\dontrun{
## Way1: Based on an annotation file in gtf/gff3 format
gtf.data <- useGff(gff = "Arabidopsis_thaliana.TAIR10.49.gtf")

##Way2: Based on a TxDb object generated from BioMart.
# Parse Arabidopsis Txdb
library(TxDb.Athaliana.BioMart.plantsmart28)
txdb <- useGff(TxDb.Athaliana.BioMart.plantsmart28)

# Parse mm10 Txdb
BiocManager::install("TxDb.Mmusculus.UCSC.mm10.ensGene")
library(TxDb.Mmusculus.UCSC.mm10.ensGene)
txdb <- useGff(TxDb.Mmusculus.UCSC.mm10.ensGene)

## Not run:
# to download gff file of a species from other databases,
# users could use biomartr::getGFF(), for example,
# download the annotation of Arabidopsis thaliana from refseq
# and store the corresponding genome file in '_ncbi_downloads/annotation'
getGFF( db       = "refseq",
        organism = "Arabidopsis thaliana",
        gunzip = TRUE
        path = file.path("_downloads","annotation"))
# then users could use useGff to load the gff file into vizAPA
useGff(gff='_downloads/annotation/Arabidopsis_thaliana_genomic_refseq.gff')
}
}
\seealso{
Other annotation functions: 
\code{\link{addAnno}()},
\code{\link{annoHub-class}},
\code{\link{getAnnoChrs}()},
\code{\link{getAnnoGenes}()},
\code{\link{getChrs}()},
\code{\link{getGenesRange}()},
\code{\link{isChrConsistent}()},
\code{\link{setDefaultAnno}()}
}
\concept{annotation functions}
