% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\docType{package}
\name{vizAPA-package}
\alias{vizAPA}
\alias{vizAPA-package}
\title{vizAPA: visualization of 3’UTR alternative polyadenylation with scRNA-seq data}
\description{
The vizAPA package is a comprehensive package for visualization of dynamics of alternative polyadenylation in single cells.
}
\details{
The vizAPA package is a comprehensive package for Visualization of dynamics of alternative PolyAdenylation in single cells.
vizTracks to plot gene model, pAs and BAM coverages, vizStats to summarize pA usages across cell categories, vizUMAP to plot 2D-embeddings, vizAPAmarkers to visualize APA markers across cell categories, etc.
The package is not species-specific, so it can be used for any species. The genome annotation could be retrieved from several sources, including gff3/gtf file, TxDb, EnsDb, BioMart, and OrganismDb. Users can provide one or more annotation sources.
}
\references{
Zhu, S., et al. (2020) PlantAPAdb: A Comprehensive Database for Alternative Polyadenylation Sites in Plants, Plant Physiol., 182, 228-242.

Wu, X., et al. (2011) Genome-wide landscape of polyadenylation in Arabidopsis provides evidence for extensive alternative polyadenylation, Proc. Natl. Acad. Sci. USA, 108, 12533-12538.
}
