% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{vizAPAMarkers}
\alias{vizAPAMarkers}
\title{Visualize APA markers}
\usage{
vizAPAMarkers(
  PACds,
  group,
  selGroups = NULL,
  markers,
  figType = c("violin", "heatmap", "bubble", "umap"),
  umap.x = NULL,
  umap.y = NULL,
  annoUMAP = TRUE,
  statTheme = NULL
)
}
\arguments{
\item{PACds}{A dataset of the PACdataset class, with each row is gene or pA.}

\item{group}{The column recording cell categories in PACds@colData.}

\item{selGroups}{A vector to specify the order and categories to plot, otherwise plot all or the first 10 (if N>10) categories.}

\item{markers}{A vector of gene names or PA ids considered as markers that should be all present in PACds@anno.}

\item{figType}{violin/heatmap/bubble/UMAP}

\item{umap.x}{The column name in PACds@colData denoting the X-axis coordinates for UMAP plot.}

\item{umap.y}{Same as umap.y but for the Y-axis.  Both umap.x and umap.y are required to be in colData. See reduceDim() for getting 2D embeddings.}

\item{annoUMAP}{Applicable when figType=umap, specify whether to plot the UMAP only with cell annotation but no expression overlay.}

\item{statTheme}{A statTHEME-like list to specify the theme of the plot.}
}
\value{
A plot.
}
\description{
vizAPAMarkers plots a violin plot, heatmap, bubble plot, or UMAP plot to show APA markers
}
\examples{
\dontrun{
data(scPACds); PACds=scPACds
m=getAPAmarkers(scPACds,  group='celltype', everyPair = TRUE)
vizAPAMarkers(PACds, group='celltype', markers=m$rowid[1:10],
             figType="violin")
vizAPAMarkers(PACds, group='celltype', markers=m$rowid[1:10],
             figType="violin", statTheme=list(xgroup=FALSE))
vizAPAMarkers(PACds, group='celltype', markers=m$rowid[1],
             figType="heatmap")
vizAPAMarkers(PACds, group='celltype', markers=m$rowid[1:10],
             figType="bubble")
vizAPAMarkers(PACds, group='celltype', selGroups=c('ES','SC'),
             markers=m$rowid[1:10], figType="bubble")
vizAPAMarkers(PACds, group='celltype', markers=m$rowid[1:10],
             figType="umap", umap.x='tsn1', umap.y='tsn2')
vizAPAMarkers(PACds, group='celltype', markers=m$rowid[1:10],
             figType="umap", umap.x='tsn1', umap.y='tsn2', annoUMAP=F)
iPACds=getAPAindexPACds(scPACds, choose2PA='PD')
m=getAPAmarkers(iPACds,  group='celltype', everyPair = TRUE, choose2PA='PD')
vizAPAMarkers(iPACds, group='celltype',
             markers=m$rowid[1:10], figType = 'violin')
vizAPAMarkers(iPACds, group='celltype', markers=m$rowid[1:10],
             figType = 'violin', statTheme=list(xgroup=FALSE))
vizAPAMarkers(iPACds, group='celltype', markers=m$rowid[1:10],
             figType = 'heatmap')
vizAPAMarkers(iPACds, group='celltype', markers=m$rowid[1:10],
             figType = 'bubble')
vizAPAMarkers(iPACds, group='celltype', markers=m$rowid[1:10],
             figType = 'bubble', statTheme=list(xgroup=FALSE))
}
}
\seealso{
Other vizMarkers functions: 
\code{\link{APAindex2PACds}()},
\code{\link{getAPAindexPACds}()},
\code{\link{getAPAmarkers}()}
}
\concept{vizMarkers functions}
