% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{vizStats}
\alias{vizStats}
\title{Draw statistic plots for bulk or single-cell PACdataset}
\usage{
vizStats(
  PACds,
  group,
  selGroups = NULL,
  gene = NULL,
  PAs = NULL,
  figType = c("violin", "box", "dot", "bubble", "heatmap"),
  log = FALSE,
  statTheme = NULL
)
}
\arguments{
\item{PACds}{A dataset of the PACdataset class.}

\item{group}{The column recording cell categories in PACds@colData.}

\item{selGroups}{A vector to specify the order and categories to plot, otherwise plot all categories.}

\item{gene}{A gene name, gene symbol, or gene id to filter PACds and gene model in the given gene.}

\item{PAs}{A vector of pA ids corresponding to rownames of PACds@counts.}

\item{figType}{figure type.
\itemize{
\item{box}: boxplot;
\item{violin}: violin plot;
\item{dot}: violin plot with dots;
\item{bubble}: bubble plot;
\item{heatmap}: heatmap which is used for bulk presentation, with each row per gene or pA and each column per sample.
If there are more than 10 columns (e.g., >10 cells), then the PACds is considered as single-cell data.
In this case, the PACds will be group and averaged by sample (e.g., merge cells to cell types), and then used for the heamtap.
}}

\item{log}{Only applicable when PACds is of count type rather then ratio type.}

\item{statTheme}{A statTHEME-like list to specify the theme of the plot.}
}
\value{
A ggplot2 plot. The x-axis is normally the cell groups (e.g., cell types), and cell groups can be selected and sorted by `selGroups`.
        The y-axis is the value or mean of (all genes or pAs if both PA and gene are not provided; one gene -- if gene is provided; individual PAs -- if PAs is provided) in all cells of each cell type.
        The order of the group is based on the factor levels of PACds@colData's group.
        If PACds@supp$row=PA, there can be multiple pAs after subsetting PACds by gene or PAs.
        In this case, there can be >1 resulting rows, which means that a cell type's boxplot will have multiple groups denoting multiple PAs.
        In other cases, the plot only shows one gene or one PA.
}
\description{
vizStats draws different types of plots, including boxplot, violin plot, dot plot, bubble plot, and heatmap, to show expression (in count or ratio) of given pAs or pAs in a gene across different conditions (e.g., cell types).
}
\examples{
\dontrun{
#plot a violin plot to show the expression levels of all pAs in a given gene across cell types.
gene=252868
vizStats(PACds, group='celltype', gene=gene, PAs=NULL, figType="violin")
#Plot other types of plots.
# boxplot
vizStats(PACds, group='celltype', gene=gene, PAs=NULL, figType="box")
# violin plot with dots
vizStats(PACds, group='celltype', gene=gene, PAs=NULL, figType="dot")
# bubble plot
vizStats(PACds, group='celltype', gene=gene, PAs=NULL, figType="bubble")
# plot the given pAs, by specifying the rowid of PAs in the PACds.
# First we get the PA ids of this gene
PAids=rownames(subsetPACds(PACds, genes=gene)@anno)
# Plot a violin plot with dots
vizStats(PACds, group='celltype', PAs=PAids, figType="dot")

# We can modify the display of the figure by changing colors or
# other parameters, providing the `statTheme` parameter.
# To use another color palette
vizStats(PACds, group='celltype', gene=gene, figType="violin",
        statTheme=list(group.cols=c(brewer.pal(8, "Set1"))))

# We can change the order of groups (e.g., cell types)
# by specifying `selGroups`.
# change the order to SC>RS>ES
vizStats(PACds, group='celltype', selGroups=c('SC','RS','ES'))

# We can also change the order of groups (e.g., cell types)
# by changing the order of the factor
# change the order to SC>RS>ES
PACds@colData$celltype=factor(PACds@colData$celltype, levels=c('SC','RS','ES'))
vizStats(PACds, group='celltype', figType="dot")

# If the expression level in PACds is count,
# we can also show the value in log scale if `log=TRUE`.
# vioin plot with dots
vizStats(PACds, group='celltype', figType="dot", log=TRUE)
# bubble plot
vizStats(PACds, group='celltype', figType="bubble", log=TRUE)

# heatmap -- for bulk
vizStats(scPACds, group='celltype', gene=gene, PAs=NULL, figType="heatmap",
         statTheme = list(heat.text.size=5))
}
}
\seealso{
Other vizStats functions: 
\code{\link{setStatTheme}()}
}
\concept{vizStats functions}
