% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{vizTracks}
\alias{vizTracks}
\title{Get all tracks for a genomic region, a gene, or pAs}
\usage{
vizTracks(
  genomicRegion = NULL,
  gene = NULL,
  PAs = NULL,
  bams = NULL,
  PACds.list = NULL,
  PA.show = c("pos"),
  cells = FALSE,
  cells.group = NULL,
  cells.width = 100,
  cells.annoCoord = NULL,
  cells.method = c("sum", "diffusion"),
  cells.sort = c("none", "all", "group"),
  annoSource = new("annoHub"),
  PA.columns = "coord",
  PA.width = 10,
  logPA = FALSE,
  space5 = 1000,
  space3 = 1000,
  vizTheme = NULL,
  res = "plot"
)
}
\arguments{
\item{genomicRegion}{A string with four fields describing a genomic region, like 1:-:1000:100000.}

\item{gene}{A gene name, gene symbol, or gene id to filter PACds and gene model in the given gene.}

\item{PAs}{A vector of pA ids corresponding to rownames of PACds@counts.}

\item{bams}{A dataframe describing the BAM file list. See the output of readBAMFileNames().}

\item{PACds.list}{A PACdataset or a list of PACdatasets.}

\item{PA.show}{A vector describing the information of pA to extract, e.g., PA.show=c('pos','tag').
PA.show=pos means extracting the coordinates of pAs; other values (e.g., tag, tot_tagnum) should be the column in PACds@anno or PACds@counts denoting the expression level of pAs.}

\item{cells}{TRUE to add track of individual cells. Parameters starting with "cells" are applicable only when cells=TRUE.}

\item{cells.group}{The column name recording cell categories in PACds@colData.}

\item{cells.width}{Similar to PA.width, but for the width of the cells bar. The final bar width=2*cell.width+1.}

\item{cells.annoCoord}{The position to put the annotation bar of cell groups, only used when group is not NULL.}

\item{cells.method}{To set the method for ordering cells, diffusion (default, using diffusion map to order cells -- see Millefy Package) or sum (using the sum of all pAs)}

\item{cells.sort}{To sort cells within group (group) or without group (all), or not sort (none, default).}

\item{annoSource}{An annoHub object.}

\item{PA.columns}{A string denoting the pA coordinates (e.g., coord) or ranges (e.g., UPA_start:UPA_end).}

\item{PA.width}{If PA.columns is a position (e.g., coord), then the pA range can be expanded by PA.width to be visualized more clearly in the track plot.
For example, if PA.width=10, then the pA position would be converted to a range of 2*PA.width+1 with the pA in the center.}

\item{logPA}{TRUE to log2(PACds@counts+1) for count-type PACds, which will affect the pA and cells tracks.}

\item{space5}{e.g, 1000 means to expand the 5' end of the gene model range by 1000 bp.}

\item{space3}{e.g, 1000 means to expand the 3' end of the gene model range by 1000 bp.}

\item{vizTheme}{a vizTHEME-like list to specify the theme of the track.}

\item{res}{A character of a PDF file name or 'plot' to plot in the current device or 'list/track' to return a list of tracks.}
}
\value{
A plot or pdf (res=plot/pdffile) or a track lists (res=track/list). The height of the track(s) of pA position is half of other tracks.
        The number of tracks depends on vizTheme (for BAM tracks), and PACds.list and PA.show (for PA tracks).
}
\description{
vizTracks gets all tracks for a genomic region, a gene, or pAs, including gene model track, pA track(s), cells track, and BAM track(s).
\itemize{
\item{"gene model track"}: plot gene models given a specific region or a gene symbol/ID, according to the annotation(s) in the annoHub object.
\item{"pA track"}: plot the positions, expression levels, APA ratios for a PACdataset or multiple PACdataset objects.
\item{"cells track"}: plot counts or ratios of individual cells using gradient colors.
\item{"BAM track"}: plot BAM coverage using lines or areas for individual cell groups or merged cell groups.
}
}
\examples{
# This is the wrapper function for visualizing the track plots,
# please see the vignette of vizAPA for full examples.
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackCells}()},
\code{\link{getTrackGeneModel}()},
\code{\link{getTrackPACds}()},
\code{\link{readBAMFileNames}()},
\code{\link{setVizTheme}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracksOfRegion}()}
}
\concept{vizTracks functions}
