% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{vizTracksOfRegion}
\alias{vizTracksOfRegion}
\title{Get all tracks in a genomic region}
\usage{
vizTracksOfRegion(
  genomicRegion,
  annoSource = new("annoHub"),
  bams = NULL,
  PACds.list = NULL,
  PA.show = c("pos"),
  cells = FALSE,
  cells.group = NULL,
  cells.width = 100,
  cells.annoCoord = NULL,
  cells.method = c("sum", "diffusion"),
  cells.sort = c("none", "all", "group"),
  PA.columns = "coord",
  PA.width = 10,
  vizTheme = NULL
)
}
\arguments{
\item{genomicRegion}{A string with four fields describing a genomic region, like 1:-:1000:100000.}

\item{annoSource}{An annoHub object. If it is empty, then plot the chromosome region instead of gene model.}

\item{bams}{A dataframe describing the BAM file list. See the output of readBAMFileNames().}

\item{PACds.list}{A PACdataset or a list of PACdatasets.}

\item{PA.show}{A vector describing the information of pA to extract, e.g., PA.show=c('pos','tag').
PA.show=pos means extracting the coordinates of pAs; other values (e.g., tag, tot_tagnum) should be the column in PACds@anno or PACds@counts denoting the expression level of pAs.}

\item{cells}{TRUE to add track of individual cells. Parameters starting with "cells" are applicable only when cells=TRUE.}

\item{cells.group}{The column name recording cell categories in PACds@colData.}

\item{cells.width}{Similar to PA.width, but for the width of the cells bar. The final bar width=2*cell.width+1.}

\item{cells.annoCoord}{The position to put the annotation bar of cell groups, only used when group is not NULL.}

\item{cells.method}{To set the method for ordering cells, diffusion (default, using diffusion map to order cells -- see Millefy Package) or sum (using the sum of all pAs)}

\item{cells.sort}{To sort cells within group (group) or without group (all), or not sort (none, default).}

\item{PA.columns}{A string denoting the pA coordinates (e.g., coord) or ranges (e.g., UPA_start:UPA_end).}

\item{PA.width}{If PA.columns is a position (e.g., coord), then the pA range can be expanded by PA.width to be visualized more clearly in the track plot.
For example, if PA.width=10, then the pA position would be converted to a range of 2*PA.width+1 with the pA in the center.}

\item{vizTheme}{a vizTHEME-like list to specify the theme of the track.}
}
\value{
A list of two elements: tracks for track list of gene model track, pA track(s) and BAM track(s); and h for track heights.
        The number of tracks depends on vizTheme (for BAM tracks), and PACds.list and PA.show (for PA tracks).
}
\description{
vizTracksOfRegion gets all tracks in a genomic region, including gene model track, pA track(s) and BAM track(s).
}
\examples{
\dontrun{
tks=vizTracksOfRegion(genomicRegion='1:+:1000:10000',
                            annoSource=new("annoHub"),
                            bams=bams, PACds.list=list(PACds, p2=PACds),
                            PA.show=list("pos","value"),
                            PA.columns="coord", PA.width=10,
                            vizTheme=NULL, toPlot=TRUE)
tracks(tks)

vizTheme=list(bam.covStyle='curve', bam.showMerge=TRUE, bam.covMerge='avg')
tks=vizTracksOfRegion(genomicRegion='1:+:1000:10000',
                  annoSource=new("annoHub"),
                  bams=bams, PACds.list=list(PACds, p2=PACds),
                  PA.show=list("pos","value"),
                  PA.columns="coord", PA.width=10,
                  vizTheme=vizTheme, toPlot=TRUE)
}
}
\seealso{
Other vizTracks functions: 
\code{\link{getTrackBams}()},
\code{\link{getTrackCells}()},
\code{\link{getTrackGeneModel}()},
\code{\link{getTrackPACds}()},
\code{\link{readBAMFileNames}()},
\code{\link{setVizTheme}()},
\code{\link{subsetPACds2GR}()},
\code{\link{vizTracks}()}
}
\concept{vizTracks functions}
