% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{win2linuxPath}
\alias{win2linuxPath}
\title{win2linuxPath converts Windows file name to Linux.}
\usage{
win2linuxPath(aPath, addBar = FALSE)
}
\arguments{
\item{aPath}{a string}

\item{addBar}{whether to add / in the end of the string, default is FALSE}
}
\value{
a new string in Linux format
}
\description{
win2linuxPath converts Windows file name to Linux.
}
\examples{
win2linuxPath('data\\\\a.bam') # "data/a.bam"
win2linuxPath('data/a.bam') # "data/a.bam"
win2linuxPath('data/a.bam', addBar=T) # "data/a.bam/"
win2linuxPath('data/a.bam/', addBar=T) # "data/a.bam/"
}
