% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizAPA_funclib.R
\name{setStatTheme}
\alias{setStatTheme}
\title{Set the plot theme for vizStats}
\usage{
setStatTheme(statTheme, check = TRUE)
}
\arguments{
\item{statTheme}{A list similar to statTHEME, containing the following parameters.
\itemize{
\item{group.cols}: parameter for vizStats or vizAPAMarkers (box/violin/point), colors for cell groups, default is `c(brewer.pal(8, "Set2"), brewer.pal(8, "Set1"))`.
\item{outlier.colour}: parameter for vizStats(box/violin/point), color for outliers, defaut is "grey".
\item{outlier.shape}: parameter for vizStats(box/violin/point), shape for outliers, defaut is 8.
\item{outlier.size}: parameter for vizStats(box/violin/point), size for outliers, default is 1.
\item{alpha}: parameter for vizStats(box/violin/point), the transparent alpha, default is 0.8.

\item{scale.low.col}: parameter for bubble/markers-dot/markers-heatmap, low point of the gradients, default is '#8d9fbb'.
\item{scale.high.col}: parameter for bubble/markers-dot/markers-heatmap, high point of the gradients, default is '#aa3e53',

\item{heat.cols}: parameter for bulk-heatmap, colors for the heatmap, default is `rev(grDevices::heat.colors(50))`.
\item{heat.text.size}: parameter for bulk-heatmap, text size, default is 5.
\item{heat.text.digits}: parameter for bulk-heatmap, number of digits for ratio value, default is 2.

\item{umap.point.size}: parameter for the UMAP plot, the point size, default is 1.
\item{umap.legend.size}: parameter for the UMAP plot, the legend text size, default is 6.
\item{umap.legend.key.size}: parameter for the UMAP plot, the legend key size, default is 0.4.
\item{umap.legend.point.size}: parameter for the UMAP plot, the legend point size, default is 1.5.

\item{xgroup}: parameter for the marker plots (violin/dot plots), whether show cell groups on X-axis, default is TRUE.
\item{hm.disp.min}: parameter of the min value for the marker plot heatmap, default is NULL (not set).
\item{hm.disp.max}: parameter of the max value for the marker plot heatmap, default is NULL (not set).
\item{hm.cluster.markers}: whether to cluster markers (rows) for the marker plot heatmap, default is FALSE.
\item{hm.cluster.dend}: whether to show dendrograms for the marker plot heatmap, default is FALSE.

\item{umaps.ncol}: parameter for the markers' multiple UMAP plots, which set the number of columns in the plot, default is 2.
}}

\item{check}{Whether to check the validity of values (default is TRUE).}
}
\value{
A list.
}
\description{
setStatTheme sets the drawing theme for vizStats, by setting default values using statTHEME and checking the validity.
}
\examples{
setStatTheme(NULL)
setStatTheme(list(group.cols=brewer.pal(8, "Set1")), check=TRUE)
}
\seealso{
Other vizStats functions: 
\code{\link{vizStats}()}
}
\concept{vizStats functions}
